PipelineController divided on Compress and decompress
